Review Rating data type update
double-byte

+++ src/Core/LibraryApp.Application/Feauters/Reviews/Commands/Create/CreateReviewCommand.cs
@@ -4,7 +4,7 @@ namespace LibraryApp.Application.Feauters.Reviews.Commands.Create;
 
     public record CreateReviewCommand(
         Guid BookId,
-        double Rating,
+        byte Rating,
         string? Title,
         string? Text) 
         : IRequest<Guid>;
